+++ ptrace/debugger/breakpoint.py
@@ -4,7 +4,7 @@ from logging import info
 from weakref import ref
 from ptrace.cpu_info import CPU_POWERPC, CPU_WORD_SIZE
 from ptrace.ctypes_tools import word2bytes
-from ptrace.six import b
+from six import b
 
 
 class Breakpoint(object):