+++ ptrace/debugger/process.py
@@ -28,7 +28,7 @@ from ptrace.debugger.process_error import ProcessError
 from ptrace.debugger.memory_mapping import readProcessMappings
 from ptrace.binding.cpu import CPU_INSTR_POINTER, CPU_STACK_POINTER, CPU_FRAME_POINTER, CPU_SUB_REGISTERS
 from ptrace.debugger.syscall_state import SyscallState
-from ptrace.six import b
+from six import b
 
 if HAS_PTRACE_SINGLESTEP:
     from ptrace.binding import ptrace_singlestep